⚙️ API/Todo - deleteTodo updated for todoIndex edit. deleteTodo - deleteTodo now update TodoIndex with new one except todo that will be deleted

+++ src/server/api/routers/todoAPI/deleteTodo.ts
@@ -1,6 +1,7 @@
 import { protectedProcedure } from "../../trpc";
 import { z } from "zod";
 import { TRPCError } from "@trpc/server";
+import { type Prisma } from "@prisma/client";
 
 // auth
 import tokenVerify from "../auth/tokenVerify";
@@ -10,6 +11,7 @@ const deleteTodo = protectedProcedure
     z.object({
       data: z.object({
         id: z.number(),
+        dateRecordId: z.number(),
       }),
     })
   )
@@ -20,7 +22,29 @@ const deleteTodo = protectedProcedure
       throw new TRPCError({ message: "TOKEN ERROR", code: "UNAUTHORIZED" });
     }
 
-    const { id } = input.data;
+    const { id, dateRecordId } = input.data;
+
+    // delete a todo dateRecordId's todo's todosIndex
+    const dateRecord = await ctx.prisma.dateRecord.findUniqueOrThrow({
+      where: { id: dateRecordId },
+      select: { todoIndex: true },
+    });
+
+    const currentTodoIndex = dateRecord.todoIndex as Prisma.JsonArray;
+    const newToodIndex = currentTodoIndex.filter(
+      (index) => index !== id
+    ) as Prisma.JsonArray;
+
+    const newDateRecord = await ctx.prisma.dateRecord.update({
+      where: { id: dateRecordId },
+      data: { todoIndex: newToodIndex },
+    });
+
+    if (!newDateRecord)
+      throw new TRPCError({
+        message: "UPDATE FAILED",
+        code: "INTERNAL_SERVER_ERROR",
+      });
 
     // RecordNotFound error when id is not correct
     const todo = await ctx.prisma.todo.delete({
